stremlit-hosting-web-scraping/requirements.txt
streamlit
gspread
pandas
google-auth
google-auth-oauthlib
matplotlib
requests
beautifulsoup4
plotly
seaborn
geopandas
